form naming changed to choreform, updated in related files & deletions made as required
